+++ packages/_private/react-demo/src/DemoForm.tsx
@@ -1,6 +1,9 @@
 /* eslint-disable no-console */
 import { useEffect, useRef, useState } from 'react';
+import { serialize } from '@synergy-design-system/components';
 import type {
+  SynCheckbox as NativeCheckbox,
+  SynFile as NativeFile,
   SynChangeEvent,
   SynRange as SynRangeNativeElement,
 } from '@synergy-design-system/components';
@@ -8,6 +11,7 @@ import {
   SynButton,
   SynCheckbox,
   SynDivider,
+  SynFile,
   SynInput,
   SynOptgroup,
   SynOption,
@@ -20,7 +24,6 @@ import {
   SynTextarea,
 } from '@synergy-design-system/react';
 import { DemoFieldset } from './DemoFieldset';
-import { normalizeData } from './shared';
 
 type FormEnabledElements = HTMLElement & {
   checked?: boolean;
@@ -34,6 +37,7 @@ const initialFormData = {
   date: '',
   donations: '2000 4000',
   email: '',
+  files: undefined,
   gender: '',
   happiness: '5',
   name: '',
@@ -69,18 +73,22 @@ export const DemoForm = () => {
     const listener = (e: SynChangeEvent) => {
       const form = formRef.current as HTMLFormElement;
 
-      const normalizedData = normalizeData(new FormData(form));
+      const normalizedData = serialize(form);
 
       // Log the normalized data
       console.log(normalizedData);
 
       // Set the field into state
       const element = e.target as FormEnabledElements;
-      const { checked, name, value } = element;
+      const { name, value } = element;
 
-      const finalValue = typeof checked !== 'undefined'
-        ? checked
-        : value;
+      let finalValue;
+
+      switch (element.tagName.toLocaleLowerCase()) {
+      case 'syn-checkbox': finalValue = (element as NativeCheckbox).checked; break;
+      case 'syn-file': finalValue = (element as NativeFile).files; break;
+      default: finalValue = value;
+      }
 
       setFormData((curr) => ({
         ...curr,
@@ -96,6 +104,8 @@ export const DemoForm = () => {
 
   return (
     <form
+      encType="multipart/form-data"
+      method="post"
       onReset={() => {
         setFormData(initialFormData);
       }}
@@ -345,6 +355,16 @@ export const DemoForm = () => {
         >
           {formData.comment}
         </SynTextarea>
+        <SynFile
+          accept="image/*"
+          droparea
+          help-text="Please upload images only"
+          files={formData.files}
+          id="screenshot"
+          label="Optional Screenshot(s)"
+          multiple
+          name="files"
+        />
       </DemoFieldset>
 
       { /* /AdditionalInformation */ }