+++ packages/_private/vanilla-demo/src/main.ts
@@ -12,7 +12,7 @@ import '@synergy-design-system/tokens/themes/light.css';
 import '@synergy-design-system/components/index.css';
 import '@synergy-design-system/styles';
 import './app.css';
-import '@synergy-design-system/components';
+import { serialize } from '@synergy-design-system/components';
 
 const capitalize = (s: string) => s && s[0].toUpperCase() + s.slice(1);
 
@@ -27,28 +27,6 @@ const setActiveView = (pathname: string) => {
     item.current = current;
   });
 };
-
-// We have to adjust the original form data
-// so it is possible to send multiple values
-// This prevents errors with select[multiple]
-// that will send all three selected elements
-// in form-data with the same name.
-// This will make sure multiple uses "," as separator for multi values
-const normalizeData = (data: FormData) => Array
-  .from(data)
-  // eslint-disable-next-line @typescript-eslint/no-base-to-string
-  .map(([key, value]) => [key, value.toString()])
-  .reduce((prev, curr) => {
-    const [currKey, currVal] = curr;
-    const finalValue = typeof prev[currKey] !== 'undefined'
-      ? `${prev[currKey]},${currVal}`
-      : currVal;
-
-    return {
-      ...prev,
-      [currKey]: finalValue,
-    };
-  }, {} as Record<string, string>);
 
 const initRouting = async () => {
   history.listen(({ location }) => {
@@ -108,7 +86,7 @@ const initApp = async () => {
   const form = document.querySelector('form')!;
 
   form.addEventListener('syn-change', () => {
-    console.log(normalizeData(new FormData(form)));
+    console.log(serialize(form));
   });
 
   form.addEventListener('submit', (e: SubmitEvent) => {