+++ packages/components/scripts/jobs/angular/createFormsModule.js
@@ -29,6 +29,13 @@ const defaultSelectors = [
 ].sort();
 
 /**
+ * List of file upload enabled elements
+ */
+const fileSelectors = [
+  'syn-file',
+];
+
+/**
  * List of components that use the checked property in forms
  * @var {string[]} checkedSelectors
  */
@@ -45,8 +52,7 @@ const checkedSelectors = [
 const createSelectors = selectors => selectors
   .map(selector => fields
     .map(field => `${selector}[${field}]`)
-    .join(', '),
-  )
+    .join(', '))
   .flat()
   .join(',\n    ');
 
@@ -98,14 +104,30 @@ export class SynDefaultValueAccessor extends DefaultValueAccessor { }
 })
 export class SynCheckedValueAccessor extends CheckboxControlValueAccessor { }
 
+@Directive({
+  providers: [{
+    multi: true,
+    provide: NG_VALUE_ACCESSOR,
+    useExisting: forwardRef(() => SynFileValueAccessor),
+  }],
+  selector: \`${createSelectors(fileSelectors)}\`,
+  // Overwrite the change event, because we only emit syn-change event
+  host: {
+    '(syn-change)': 'onChange($event.target.files)',
+  },
+})
+export class SynFileValueAccessor extends DefaultValueAccessor { }
+
 @NgModule({
   declarations: [
     SynDefaultValueAccessor,
     SynCheckedValueAccessor,
+    SynFileValueAccessor,
   ],
   exports: [
     SynDefaultValueAccessor,
     SynCheckedValueAccessor,
+    SynFileValueAccessor,
   ],
   imports: [],
 })