+++ packages/components/src/translations/en.ts
@@ -45,8 +45,15 @@ const translation: Translation = {
   // @see scripts/vendorism/translations.vendorism.js
   closeMenu: 'Close menu',
   danger: 'Danger',
+  fileButtonText: 'Choose file',
+  fileButtonTextMultiple: 'Choose files',
+  fileDragDrop: 'Drop or choose file',
   menu: 'Menu',
   notification: 'Notification',
+  numFilesSelected: num => {
+      if (num === 0) return 'No files chosen';
+      return `${num} files`;
+    },
   openMenu: 'Open menu',
   rangeMax: 'Maximum',
   rangeMin: 'Minimum',